CoupledNODE

Documentation for CoupledNODE.

CoupledNODE.lhist

CoupledNODE.Grid

CoupledNODE.Concatenate

CoupledNODE.Unpack

CoupledNODE.create_f_CNODE

CoupledNODE.create_randloss_MulDtO

CoupledNODE.create_randloss_derivative

CoupledNODE.loss_MulDtO_oneset

CoupledNODE.mean_squared_error

CoupledNODE.predict_u_CNODE

CoupledNODE.lhist — Constant

callback(p, l, pred; doplot=true)

The callback function is used to observe training progress. It prints the current loss value and updates the rolling average of the loss. If do_plot is true, it also plots the rolling average of the loss every 10 steps.

Arguments

p: the current parameters

l: the current loss value

pred: the current prediction

do_plot: a boolean indicating whether to plot the rolling average of the loss every 10 steps (default is true)

Returns

false if nothing unexpected happened.

CoupledNODE.Grid — Type

struct Grid

Object containing the grid information. It can handle 1D, 2D, and 3D grids.

Fields:

dim::Int: The dimensionality of the grid.

dx::Union{Float32, Float64}: The grid spacing in the x-direction.

dy::Union{Float32, Float64}: The grid spacing in the y-direction. Default is 0 for 1D grid.

dz::Union{Float32, Float64}: The grid spacing in the z-direction. Default is 0 for 1D and 2D grids.

nx::Int: The number of grid points in the x-direction.

ny::Int: The number of grid points in the y-direction. Default is 0 for 1D grid.

nz::Int: The number of grid points in the z-direction. Default is 0 for 1D and 2D grids.

N::Int: The total number of elements in the grid.

Constructor:

Grid(dim::Int, dx::Union{Float32, Float64}, dy::Union{Float32, Float64} = 0, dz::Union{Float32, Float64} = 0, nx::Int, ny::Int = 0, nz::Int = 0, convert_to_float32::Bool = false): Constructs a Grid object with the given grid parameters. The dy, dz, ny, and nz parameters can be omitted for 1D and 2D grids.

CoupledNODE.Concatenate — Method

Concatenate(grids)

Creates a function to concatenate the coupled variables to a single vector.

Arguments

grids: A vector or tuple containing the grid(s).

Returns

A list of concatenated coupled variables.

CoupledNODE.Unpack — Method

Unpack(grids)

Creates a function to unpack the input data from a concatenated list to a tuple.

Arguments

grids: A vector or tuple containing the grid(s).

Returns

A list of the unpacked data (coupled variables)

CoupledNODE.create_f_CNODE — Function

create_f_CNODE(forces, grids, NNs = nothing; pre_force = identity,
    post_force = identity, is_closed = false)

Create a CoupledNODE (CNODE) function that represents a system of coupled differential equations.

Arguments

forces: A vector or tuple of functions representing the forces in the system.

grids: A vector or tuple of grids representing the variables in the system.

NNs: (optional) A vector or tuple of neural networks representing the closure terms in the system. Default is nothing.

pre_force: (optional) A function to be applied before the force layer. Default is identity.

post_force: (optional) A function to be applied after the force layer. Default is identity.

is_closed: (optional) A boolean indicating whether the CNODE is closed. Default is false.

Returns

A Chain object representing the consecutive set of operations taking place in the CNODE.

CoupledNODE.create_randloss_MulDtO — Method

create_randloss_MulDtO(target; nunroll, nintervals=1, nsamples, λ_c, λ_l1)

Creates a random loss function for the multishooting method with multiple shooting intervals.

Arguments

target: The target data for the loss function.

training_CNODE: Model CNODE.

st: state of the neural part.

nunroll: The number of time steps to unroll.

noverlaps: The number of time steps that overlaps between consecutive intervals.

nintervals: The number of shooting intervals.

nsamples: The number of samples to select.

λ_c: The weight for the continuity term. It sets how strongly we make the pieces match (continuity term).

λ_l1: The coefficient for the L1 regularization term in the loss function.

Returns

randloss_MulDtO: A random loss function for the multishooting method.

CoupledNODE.create_randloss_derivative — Method

create_randloss_derivative(input_data, F_target, f, st; n_use = size(input_data, 2), λ=0, λ_c = 0)

Create a randomized loss function that compares the derivatives. This is done because using the entire dataset at each iteration would be too expensive.

Arguments

input_data: The input data.

F_target: The target data.

f: The model function.

st: The model state.

n_use: The number of samples to use for the loss function. Defaults to the number of samples in the input data.

λ and λ_c: The regularization parameter. Defaults to 0.

Returns

A function that computes the mean squared error loss and takes as input the model parameters.

CoupledNODE.loss_MulDtO_oneset — Method

loss_MulDtO_oneset(trajectory, θ; λ_c=1e1, λ_l1=1e1, nunroll, nintervals, nsamples=nsamples)

Compute the loss function for the multiple shooting method with a continuous neural ODE (CNODE) model. Check https://docs.sciml.ai/DiffEqFlux/dev/examples/multiple_shooting/ for more details.

Arguments

trajectory: The trajectory of the system.

θ: The parameters of the CNODE model.

training_CNODE: Model CNODE.

λ_c: The weight for the continuity term. It sets how strongly we make the pieces match (continuity term). Default is 1e1.

λ_l1: The weight for the L1 regularization term. Default is 1e1.

nunroll: The number of time steps to unroll the trajectory.

noverlaps: The number of time steps that overlaps between each consecutive intervals.

nintervals: The number of intervals to divide the trajectory into.

nsamples: The number of samples.

Returns

loss: The computed loss value.

CoupledNODE.mean_squared_error — Method

mean_squared_error(f, st, x, y, θ, λ)

Random a priori loss function. Use this function to train the closure term to reproduce the right hand side.

Arguments:

f: Function that represents the model.

st: State of the model.

x: Input data.

y: Target data.

θ: Parameters of the model.

λ and λ_c: Regularization parameters.

Returns:

total_loss: Mean squared error loss.

CoupledNODE.predict_u_CNODE — Method

predict_u_CNODE(uv0, θ, st, nunroll, training_CNODE, tg)

Auxiliary function to solve the NeuralODE. Returns the prediction and the target sliced to nunroll time steps for convenience when calculating the loss.

Arguments

uv0: Initial condition(s) for the NeuralODE.

θ: Parameters of the NeuralODE.

st: Time steps for the NeuralODE.

nunroll: Number of time steps in the window.

training_CNODE: CNODE model that solves the NeuralODE.

tg: Target values.

Returns

sol: Prediction of the NeuralODE sliced to nunroll time steps.

tg[:, :, 1:nunroll]: Target values sliced to nunroll time steps.