CoupledNODE

Documentation for CoupledNODE.

CoupledNODE.lhist

CoupledNODE.Grid

CoupledNODE.create_f_CNODE

CoupledNODE.create_f_NODE

CoupledNODE.create_randloss_MulDtO

CoupledNODE.create_randloss_derivative

CoupledNODE.downscale_v

CoupledNODE.loss_MulDtO_oneset

CoupledNODE.mean_squared_error

CoupledNODE.upscale_v

CoupledNODE.lhist — Constant

callback(p, l, pred; doplot=true)

The callback function is used to observe training progress. It prints the current loss value and updates the rolling average of the loss. If do_plot is true, it also plots the rolling average of the loss every 10 steps.

Arguments

p: the current parameters

l: the current loss value

pred: the current prediction

do_plot: a boolean indicating whether to plot the rolling average of the loss every 10 steps (default is true)

Returns

false if nothing unexpected happened.

CoupledNODE.Grid — Type

struct Grid

This object contains the grid information.

Fields:

dux::Float64: The grid spacing in the x-direction for u.

duy::Float64: The grid spacing in the y-direction for u.

nux::Int: The number of grid points in the x-direction for u.

nuy::Int: The number of grid points in the y-direction for u.

dvx::Float64: The grid spacing in the x-direction for v.

dvy::Float64: The grid spacing in the y-direction for v.

nvx::Int: The number of grid points in the x-direction for v.

nvy::Int: The number of grid points in the y-direction for v.

Nu::Int: The total number of elements for u.

Nv::Int: The total number of elements for v.

Constructor:

Grid(dux::Float64, duy::Float64, nux::Int, nuy::Int, dvx::Float64, dvy::Float64, nvx::Int, nvy::Int): Constructs a Grid object with the given grid parameters.

CoupledNODE.create_f_CNODE — Function

create_f_CNODE(F_u, G_v, grid, NN_u=nothing, NN_v=nothing; is_closed=false)

Create a neural network model for the Coupled Neural ODE (CNODE) approach.

Arguments

F_u: Function that defines the right-hand side of the CNODE for the variable u.

G_v: Function that defines the right-hand side of the CNODE for the variable v.

grid: Grid object that represents the spatial discretization of the variables u and v.

NN_u (optional): Neural network model for the variable u. Default is nothing.

NN_v (optional): Neural network model for the variable v. Default is nothing.

is_closed (optional): Boolean indicating whether the CNODE is closed or not. Default is false.

Returns

The created CNODE model.

CoupledNODE.create_f_NODE — Method

create_f_NODE(NN, f_u; is_closed=false)

Create a Neural ODE (NODE) using ResNet skip blocks to add the closure.

Arguments

NN: The neural network model.

f_u: The closure function.

is_closed: A boolean indicating whether to add the closure or not. Default is false.

Returns

The created Neural ODE (NODE) model.

CoupledNODE.create_randloss_MulDtO — Method

create_randloss_MulDtO(target; nunroll, nintervals=1, nsamples, λ_c, λ_l1)

Creates a random loss function for the multishooting method with multiple shooting intervals.

Arguments

target: The target data for the loss function.

training_CNODE: Model CNODE.

nunroll: The number of time steps to unroll.

noverlaps: The number of time steps that overlaps between each consecutive intervals.

nintervals: The number of shooting intervals.

nsamples: The number of samples to select.

λ_c: The weight for the continuity term. It sets how strongly we make the pieces match (continuity term).

λ_l1: The coefficient for the L1 regularization term in the loss function.

Returns

randloss_MulDtO: A random loss function for the multishooting method.

CoupledNODE.create_randloss_derivative — Method

create_randloss_derivative(GS_data, FG_target, f, st; nuse = size(GS_data, 2), λ=0)

Create a randomized loss function that compares the derivatives. This function creates a randomized loss function derivative by selecting a subset of the data. This is done because using the entire dataset at each iteration would be too expensive.

Arguments

GS_data: The input data.

FG_target: The target data.

f: The model function.

st: The model state.

nuse: The number of samples to use for the loss function. Defaults to the size of GS_data.

λ: The regularization parameter. Defaults to 0.

Returns

A function randloss that computes the mean squared error loss using the selected subset of data.

CoupledNODE.downscale_v — Function

downscale_v(grid, resize_v=false)

Generate the layer that downscales v and expects as input a tuple (u, v). If resize_v is true, the layer upscales v to twice the target size and applies mean pooling.

Arguments:

grid: Grid object that represents the spatial discretization of the variables u and v.

resize_v: A boolean indicating whether to resize v.

Returns:

Function that takes a tuple (u, v) returns the linearized concatenation of u and v.

CoupledNODE.loss_MulDtO_oneset — Method

loss_MulDtO_oneset(trajectory, θ; λ_c=1e1, λ_l1=1e1, nunroll, nintervals, nsamples=nsamples)

Compute the loss function for the multiple shooting method with a continuous neural ODE (CNODE) model. Check https://docs.sciml.ai/DiffEqFlux/dev/examples/multiple_shooting/ for more details.

Arguments

trajectory: The trajectory of the system.

θ: The parameters of the CNODE model.

training_CNODE: Model CNODE.

λ_c: The weight for the continuity term. It sets how strongly we make the pieces match (continuity term). Default is 1e1.

λ_l1: The weight for the L1 regularization term. Default is 1e1.

nunroll: The number of time steps to unroll the trajectory.

noverlaps: The number of time steps that overlaps between each consecutive intervals.

nintervals: The number of intervals to divide the trajectory into.

nsamples: The number of samples. Default is nsamples.

Returns

loss: The computed loss value.

nothing: Placeholder return value.

CoupledNODE.mean_squared_error — Method

mean_squared_error(f, st, x, y, θ, λ)

Random a priori loss function. Use this function to train the closure term to reproduce the right hand side.

Arguments:

f: Function that represents the model.

st: State of the model.

x: Input data.

y: Target data.

θ: Parameters of the model.

λ: Regularization parameter.

Returns:

total_loss: Mean squared error loss.

CoupledNODE.upscale_v — Function

upscale_v(grid, resize_v=false)

Generate the layer that upscales v. This function expects as input the linearized concatenation of u and v and returns a tuple (u, v).

Arguments:

grid: Grid object that represents the spatial discretization of the variables u and v.

resize_v: A boolean indicating whether to resize the v component.

Returns:

Function that reshapes u and v on the grid and returns a tuple (u, v)